created QuantityTableDataProvider which transforms supplied harvest data into an object of counts of each individual plant type
