+++ src/scripts/quantityTable/QuantityTable.js
@@ -1,12 +1,12 @@
+import { usePlantCounts } from './QuantityTableDataProvider.js';
 import { QuantityTableRow } from './QuantityTableRow.js';
 
 export const QuantityTable = harvest => {
   const domNode = document.querySelector('.container');
 
-  const counts = getCounts(harvest);
-  const countsArray = Object.keys(counts).map(type => ({ type, count: counts[type] }));
+  const counts = usePlantCounts(harvest);
 
-  const tableRowsHTML = countsArray.map(QuantityTableRow).join('\n');
+  const tableRowsHTML = counts.map(QuantityTableRow).join('\n');
 
   domNode.innerHTML += `
     <table class="quantity-table">
@@ -22,20 +22,3 @@ export const QuantityTable = harvest => {
     </table>
   `;
 };
-
-/**
- * Given an array of plant objects, return object of the form: { [plant.type]: <count of occurences of that plant in input array> }
- * @param {Array} harvest Array of plant objects
- */
-const getCounts = harvest => {
-  const counts = {};
-
-  harvest.forEach(plant => {
-    const { type } = plant;
-
-    const count = counts[type] ? counts[type] + 1 : 1;
-    counts[type] = count;
-  });
-
-  return counts;
-}
